+++ js/client/index.js
@@ -14,6 +14,7 @@
  */
 
 export { call } from './call.js'
+export { coalesce } from './coalesce.js'
 export { callImage } from './call_image.js'
 export { callTranscription } from './call_transcription.js'
 export { callEmbedding } from './call_embedding.js'